Refactor: Split test_main.py into multiple files
Split the large test_main.py file into several smaller, more
focused test files, each containing a single test class. This
improves test organization and readability.

- Created test_emails.py for TestEmails and related helpers.
- Created test_fetch.py for TestFetch and related webserver
  functions.
- Created test_send.py for TestSend and its webserver function.
- Created test_feed_config.py for TestFeedConfig.
- Created test_opml.py for TestOPML.
- Updated test_main.py renamed to test_doctests as only includes
  doctest loading.
