+++ test/test_reply_changes.py
@@ -12,6 +12,9 @@ sys.path.insert(0, os.path.dirname(__file__))
 from util.execcontext import ExecContext
 from util.tempmaildir import TemporaryMaildir
 
+# Directory containing test feed data/configs
+test_dir = str(Path(__file__).absolute().parent.joinpath("data"))
+
 
 class TestReplyChanges(unittest.TestCase):
 
@@ -51,19 +54,21 @@ class TestReplyChanges(unittest.TestCase):
 
     def _call(self, config: str):
         with ExecContext(config) as ctx:
-            shutil.copyfile("data/nodejs/feed1.xml", str(self.feed_path))
+            shutil.copyfile(
+                os.path.join(test_dir, "nodejs/feed1.xml"), str(self.feed_path)
+            )
             ctx.call("run")
-            shutil.copyfile("data/nodejs/feed2.xml", str(self.feed_path))
+            shutil.copy(os.path.join(test_dir, "nodejs/feed2.xml"), str(self.feed_path))
             ctx.call("run")
 
     def test_switch(self):
         config = self._config({"reply-changes": False})
         with ExecContext(config) as ctx:
-            shutil.copyfile("data/nodejs/feed1.xml", str(self.feed_path))
+            shutil.copy(os.path.join(test_dir, "nodejs/feed1.xml"), str(self.feed_path))
             ctx.call("run")
             ctx.change_config({"reply-changes": True})
             ctx.call("run")
-            shutil.copyfile("data/nodejs/feed2.xml", str(self.feed_path))
+            shutil.copy(os.path.join(test_dir, "nodejs/feed2.xml"), str(self.feed_path))
             ctx.call("run")
 
         messages = self.maildir.inbox_messages()  # type: List[mailbox.MaildirMessage]
@@ -74,11 +79,11 @@ class TestReplyChanges(unittest.TestCase):
     def test_no_send(self):
         config = self._config({"reply-changes": True})
         with ExecContext(config) as ctx:
-            shutil.copyfile("data/nodejs/feed1.xml", str(self.feed_path))
+            shutil.copy(os.path.join(test_dir, "nodejs/feed1.xml"), str(self.feed_path))
             ctx.call("run")
-            shutil.copyfile("data/nodejs/feed2.xml", str(self.feed_path))
+            shutil.copy(os.path.join(test_dir, "nodejs/feed2.xml"), str(self.feed_path))
             ctx.call("run", "--no-send")
-            shutil.copyfile("data/nodejs/feed3.xml", str(self.feed_path))
+            shutil.copy(os.path.join(test_dir, "nodejs/feed3.xml"), str(self.feed_path))
             ctx.call("run")
 
         messages = self.maildir.inbox_messages()  # type: List[mailbox.MaildirMessage]
